# docs/swagger.yaml
basePath: /swagger/box-manage-service
definitions:
  controllers.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      timestamp:
        example: "2025-01-26T12:00:00Z"
        type: string
    type: object
...
  controllers.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      timestamp:
        example: "2025-01-26T12:00:00Z"
        type: string
      total:
        example: 100
        type: integer
    type: object
...
  models.VideoSource:
    properties:
      created_at:
        description: 创建时间
        example: "2025-01-26T12:00:00Z"
        type: string
      description:
        type: string
      error_msg:
        type: string
      id:
        description: 主键ID
        example: 1
        type: integer
      last_active:
        type: string
      name:
        description: 基本信息
        type: string
      password:
        type: string
      play_url:
        description: 播放地址
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.VideoSourceStatus'
        description: 状态信息
      stream_id:
        description: ZLMediaKit 流信息
        type: string
      type:
        $ref: '#/definitions/models.VideoSourceType'
      updated_at:
        description: 更新时间
        example: "2025-01-26T12:00:00Z"
        type: string
      url:
        description: 连接信息
        type: string
      user_id:
        description: 关联信息
        type: integer
      username:
        type: string
      video_file_id:
        description: 关联的视频文件ID（仅文件类型使用）
        type: integer
    type: object
  models.VideoSourceStatus:
    enum:
    - inactive
    - active
    - error
    type: string
    x-enum-comments:
      VideoSourceStatusActive: 活跃
      VideoSourceStatusError: 错误
      VideoSourceStatusInactive: 未激活
    x-enum-descriptions:
    - 未激活
    - 活跃
    - 错误
    x-enum-varnames:
    - VideoSourceStatusInactive
    - VideoSourceStatusActive
    - VideoSourceStatusError
  models.VideoSourceType:
    enum:
    - stream
    - file
    type: string
    x-enum-comments:
      VideoSourceTypeFile: 视频文件
      VideoSourceTypeStream: 实时流
    x-enum-descriptions:
    - 实时流
    - 视频文件
    x-enum-varnames:
    - VideoSourceTypeStream
    - VideoSourceTypeFile
...
  service.CreateVideoSourceRequest:
    properties:
      description:
        type: string
      name:
        type: string
      password:
        type: string
      stream_id:
        type: string
      type:
        $ref: '#/definitions/models.VideoSourceType'
      url:
        description: 流类型必需，文件类型可选
        type: string
      user_id:
        type: integer
      username:
        type: string
      video_file_id:
        description: 文件类型视频源关联的文件ID
        type: integer
    required:
    - name
    - type
    - user_id
    type: object
...
  service.ScreenshotResponse:
    properties:
      base64_data:
        description: 图片的base64编码
        type: string
      content_type:
        description: 内容类型，如image/jpeg
        type: string
      file_size:
        description: 文件大小（字节）
        type: integer
      height:
        description: 图片高度
        type: integer
      timestamp:
        description: 截图时间戳
        type: string
      width:
        description: 图片宽度
        type: integer
    type: object
...
  service.UpdateVideoSourceRequest:
    properties:
      description:
        type: string
      name:
        type: string
      password:
        type: string
      stream_id:
        type: string
      url:
        type: string
      username:
        type: string
      video_file_id:
        description: 指向指针的指针，支持设置为null
        type: integer
    type: object
info:
  contact: {}
  description: AI盒子管理系统后端服务，提供盒子管理、模型管理、任务管理、用户管理等功能
  title: AI盒子管理系统 API
  version: "1.0"
paths:
...
  /api/v1/video/sources:
    get:
      consumes:
      - application/json
      description: 分页获取视频源列表，支持按名称和类型搜索
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 视频源名称搜索关键词
        in: query
        name: name
        type: string
      - description: 视频源类型
        enum:
        - file
        - stream
        - camera
        in: query
        name: type
        type: string
      - description: 视频源状态
        enum:
        - active
        - inactive
        - error
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/controllers.PaginatedResponse'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.VideoSource'
                        type: array
                    type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取视频源列表
      tags:
      - 视频源管理
    post:
      consumes:
      - application/json
      description: 创建新的视频源
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateVideoSourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoSource'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 创建视频源
      tags:
      - 视频源管理
  /api/v1/video/sources/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的视频源
      parameters:
      - description: 视频源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 视频源不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除视频源
      tags:
      - 视频源管理
    get:
      consumes:
      - application/json
      description: 根据ID获取视频源详细信息
      parameters:
      - description: 视频源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoSource'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 视频源不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取视频源详情
      tags:
      - 视频源管理
    put:
      consumes:
      - application/json
      description: 更新视频源信息
      parameters:
      - description: 视频源ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateVideoSourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoSource'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 视频源不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 更新视频源
      tags:
      - 视频源管理
  /api/v1/video/sources/{id}/screenshot:
    post:
      consumes:
      - application/json
      description: 对指定视频源进行截图，返回base64编码的jpg图片
      parameters:
      - description: 视频源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 截图成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.ScreenshotResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 视频源不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 视频源截图
      tags:
      - 视频源管理
...
swagger: "2.0"